requirements.txt
streamlit
python-dotenv
requests
pandas
notion-client
beautifulsoup4
